--- react-order-master/src/App.js
@@ -1,9 +1,0 @@
-import React from "react"
-import './App.css';
-import MyTable from  './components/MyTable';
-
-function App() {
-  return <MyTable />;
-}
-
-export default App;
